First_Attempt
This is my first attempt of creating this finance calculator. If the user enters an invalid option (anything but investment, bond, simple or compound), the program will advise that the user has made an invalid entry and to choose from the given options. However, once the user has re-entered its choice, the program will end.