Improved_Version2
My logical error in the previous "Improved version": I was checking if the strings themselves are true instead of checking if they are equal to a string.

After realising my error, I have amended the while loops, which allows the program to break out the outer loop and execute the "else" statements if needed, which was not possible previously. 

The next improvement I would like to tackle is defensive programming/exception handling. I want to make sure that the program does not crash if the user enters a non-numeric value.