Improved_Version3
As the previous program would have crashed, as soon the user would have entered a non-numeric value for certain input requests, I have amended the code in order to handle exceptions. 
So if the user enters a non-numeric value (where the program is specifically asking for a numeric value), the user will receive an error message and be prompted to enter a numeric value. 
The program will proceed with the calculation as soon all values have been received.